# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ErrorResponse:
    properties:
      error_code:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.NotificationPreferences:
    properties:
      on_email:
        example: true
        type: boolean
      on_sms:
        example: false
        type: boolean
      user_id:
        example: cb7bc97f-45b0-4972-8edf-dc7300cc059c
        type: string
    type: object
  models.PartialUser:
    properties:
      created_at:
        example: "1977-04-22T06:00:00Z"
        type: string
      first_name:
        example: Henri
        type: string
      is_verified:
        example: true
        type: boolean
      last_name:
        example: Martin
        type: string
    type: object
  models.StripeIdentityVerificationResponse:
    properties:
      id:
        type: string
      next_action:
        properties:
          redirect_to_url:
            type: string
          type:
            type: string
        type: object
      person:
        type: string
      refresh_url:
        type: string
      return_url:
        type: string
      status:
        type: string
    type: object
  models.TimeSlot:
    properties:
      end_time:
        example: "18:00:00"
        type: string
      start_time:
        example: "14:00:00"
        type: string
      user_id:
        type: string
      weekday:
        example: 0
        type: integer
    type: object
  models.User:
    properties:
      additional_street_information:
        example: Residence des peupliers
        type: string
      birthdate:
        example: "1999-10-05"
        type: string
      city:
        example: Lille
        type: string
      civility:
        example: Mme
        type: string
      country:
        example: FR
        type: string
      email:
        example: henri.martin@qimpl.fr
        type: string
      first_name:
        example: Henri
        type: string
      is_admin:
        example: true
        type: boolean
      is_deleted:
        example: true
        type: boolean
      is_enabled:
        example: true
        type: boolean
      is_owner:
        example: true
        type: boolean
      last_name:
        example: Martin
        type: string
      mobile_phone_number:
        example: "0610890978"
        type: string
      notification_preferences:
        $ref: '#/definitions/models.NotificationPreferences'
        type: object
      password:
        example: MyPassword
        type: string
      state:
        example: Haut-De-France
        type: string
      street:
        example: Rue des peupliers
        type: string
      stripe_account_id:
        example: acct_1HqMQH2Hlu9RYi7N
        type: string
      stripe_customer_id:
        example: cus_IOwdRp9gIlOjTD
        type: string
      stripe_payment_method_id:
        example: pm_1Ho8k8CMhQMU3AqAKJwPYAXj
        type: string
      user_verifications:
        $ref: '#/definitions/models.UserVerifications'
        type: object
      zip:
        example: "59000"
        type: string
    type: object
  models.UserLogin:
    properties:
      email:
        example: henri.martin@qimpl.fr
        type: string
      password:
        example: MyPassword
        type: string
    type: object
  models.UserResetPasswordBody:
    properties:
      password:
        type: string
    type: object
  models.UserVerifications:
    properties:
      is_verified:
        example: true
        type: boolean
      stripe_person_id:
        example: vip_IRVyOgajqmYpMq
        type: string
      user_id:
        example: cb7bc97f-45b0-4972-8edf-dc7300cc059c
        type: string
      verification_type:
        example: identity_document
        type: string
      verified_at:
        type: string
    type: object
info:
  contact: {}
  license: {}
  title: authentication API
  version: 0.1.0
paths:
  /authenticate:
    post:
      description: Control user credential and generate a new Json Web Token
      parameters:
      - description: UserLogin object
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: body
        "400":
          description: Bad Request
          schema:
            type: body
        "422":
          description: Unprocessable Entity
          schema:
            type: body
      summary: Can allow user to login and get an JWT.
      tags:
      - Authentication
  /time-slots:
    post:
      consumes:
      - application/json
      description: Create a new time slot of a user
      parameters:
      - description: TimeSlot information
        in: body
        name: TimeSlot
        required: true
        schema:
          $ref: '#/definitions/models.TimeSlot'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: body
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Create a new time slot
      tags:
      - Time Slots
  /time-slots/{time_slot_id}:
    put:
      consumes:
      - application/json
      description: Update Time Slot by its ID
      parameters:
      - description: Time Slot ID
        in: query
        name: time_slot_id
        required: true
        type: string
      - description: Time Slot information
        in: body
        name: TimeSlot
        required: true
        schema:
          $ref: '#/definitions/models.TimeSlot'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: body
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Update Time Slot
      tags:
      - Time Slots
  /time-slots/user/{user_id}:
    get:
      description: Get all time slots of a user using his ID
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: body
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all time slots of a user
      tags:
      - Time Slots
  /user:
    get:
      description: Get users array objects data from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get users from database
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User object
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create user
      tags:
      - Users
  /user/{user_id}:
    delete:
      description: Delete user data from database
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete user by his id
      tags:
      - Users
    get:
      description: Get user object data from database
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user by his id
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an user
      parameters:
      - description: User object
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update User by his ID
      tags:
      - Users
  /user/{user_id}/anonymize:
    put:
      description: Anonymize user data & set is_deleted boolean to true
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: anonymize user account by his id.
      tags:
      - Users
  /user/{user_id}/desactivate:
    put:
      description: Update user is_enabled field into database
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Desactivate user account by his id.
      tags:
      - Users
  /user/{user_id}/identity/verification_intent:
    post:
      description: Create Stripe Link et insert into database a new identity verification process
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StripeIdentityVerificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create Stripe link intent to validate identity for a given user
      tags:
      - Users
  /user/{user_id}/partial:
    get:
      description: Get partial user object data from database
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PartialUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get partial user by its id
      tags:
      - Users
  /user/{user_id}/reset/password:
    put:
      description: Update user password field into database
      parameters:
      - description: UserResetPasswordBody object
        in: body
        name: UserResetPasswordBody
        required: true
        schema:
          $ref: '#/definitions/models.UserResetPasswordBody'
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Reset user password by his id
      tags:
      - Users
  /user/{user_id}/validate:
    put:
      description: Update user is_enabled field into database
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Validate user account by his id.
      tags:
      - Users
swagger: "2.0"